parlib: Check externally before running ctors
Due to how we force gcc to link parlib with every binary, our shared
objects (e.g. libelf.so) have a copy of parlib and its ctors.  These
ctors run when the shared obj is loaded.  Additionally, the ctors run
again when the binary is loaded.  This means we've been running the
ctors multiple times (for perf and vmrunkernel), and haven't been
noticing it.  Yikes!

The init_once_racy() test is only sufficient for calls from within a
particular object.  E.g. within libelf, pthread_test, etc.  So we need
to check an external location.  Procdata works for this.  If the ctors
for parlib ran, then uthread_lib_init() would have told the kernel we're
ready for vcore context.  Later parlib ctors can check that flag too.

This also means that if you change parlib, you need to rebuild any
shared objects and refill-kfs to make sure those changes propagate to
all programs.  I'm not sure of a better way around this for now.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>